requirements.txt
streamlit
joblib
numpy
pillow
scikit-learn
